+++ include/analyzer/modules/sfr/sfr_analyzer.h
@@ -1,0 +1,10 @@
+/* SFR-Analyzer Submodule Header */
+
+#ifndef SFR_ANALYZER_H
+#define SFR_ANALYZER_H
+
+typedef struct _report report_t;
+
+extern int analyze_sfr(report_t *report);
+
+#endif